Add ScummVM engine
This commit introduces the ScummVM game engine to the build system. It
includes the necessary configuration files, patches for specific
functionalities like screenshot directory handling and Raspberry Pi SDL2
configuration, and a Makefile for building the engine.

A `scummvm.keys` file is also included to provide default virtual
keyboard bindings for controller input.

Signed-off-by: Juliano Dorigão <jdorigao@gmail.com>